added the "save" packet dump to opendcre_scan, added a bit of docs
the "save" packet that gets sent over the bus upon scan success is a DumpCommand with the SAVE bit and SCAN_ALL bit enabled in the upper byte. currently this change should have no effect, since the firmware needs to be updated. the emulator should be updated to ignore DumpCommands with the SAVE bit enabled, since we don't expect a response back.