fix bug in flask endpoint opendcre_scan, update tests for recent changes
ipmi scan_all tests failed previously because the retry mechanism maxed out, as designed, when invalid data was found. since some of the configuration files intentionally contain invalid data for the tests and are static configurations per each test fixture, BusCommunicationError was thrown. The temporary fix is to add a new json configuration file which contains only valid entries within it.

in the long run, some kind of test harness should be designed which would allow for proper testing of the retry, where board/device values can change within the configuration based on the retry count.

note: this commit fixes some, but not all tests, so errors/failures still exist